flask-app/requirements.txt
dogpile.cache==0.6.4
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
pendulum==1.2.4
MarkupSafe==0.23
requests==2.8.1
Werkzeug==0.11.9
